+++ metadata-bop/src/main/java/com/mingdong/bop/service/TradeService.java
@@ -11,13 +11,13 @@ public interface TradeService
     /**
      * 根据条件获取充值记录列表
      */
-    void getProductRechargeInfoList(String keyword, Long productId, Long rechargeType, Date fromDate, Date toDate,
+    void getProductRechargeInfoList(String keyword, Long productId, Integer rechargeType, Date fromDate, Date toDate,
             Page page, RestListResp res);
 
     /**
      * 根据条件导出充值记录列表
      */
-    XSSFWorkbook createProductRechargeInfoListXlsx(String keyword, Long productId, Long managerId, Long rechargeType,
+    XSSFWorkbook createProductRechargeInfoListXlsx(String keyword, Long productId, Long managerId, Integer rechargeType,
             Date fromDate, Date toDate, Page page);
 
     void getClientBillList(String keyword, Long productId, Integer billPlan, Date fromDate, Date toDate, Long managerId,
